backend/requirements.txt
alembic==1.13.0
ansi2html==1.8.0
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.2
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
plotly==5.18.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
reportlab==4.0.8
requests==2.31.0
retrying==1.3.4
six==1.16.0
SQLAlchemy==2.0.23
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0
